+++ common/src/main/java/com/human/common/registry/tag/HumanItemTags.java
@@ -21,6 +21,8 @@ public class HumanItemTags {
 
     public static final TagKey<Item> MK50_ARMOR = create("mk50_armor");
 
+    public static final TagKey<Item> MK50_ARMOR_REPAIR_INGREDIENTS = create("mk50_armor_repair_ingredients");
+
     public static final TagKey<Item> PADDING_BLOCKS = create("padding_blocks");
 
     public static final TagKey<Item> PLASTIC = create("plastic");
@@ -29,11 +31,19 @@ public class HumanItemTags {
 
     public static final TagKey<Item> PRESSURE_ARMOR = create("pressure_armor");
 
+    public static final TagKey<Item> PRESSURE_ARMOR_REPAIR_INGREDIENTS = create("pressure_armor_repair_ingredients");
+
     public static final TagKey<Item> RADIATION_CURE_ITEMS = create("radiation_cure_items");
 
     public static final TagKey<Item> RADIATION_RESISTANT_ARMORS = create("radiation_resistant_armors");
 
     public static final TagKey<Item> RADIOACTIVE_ITEMS = create("radioactive_items");
+
+    public static final TagKey<Item> STEEL_ARMOR_REPAIR_INGREDIENTS = create("steel_armor_repair_ingredients");
+
+    public static final TagKey<Item> TACTICAL_ARMOR_REPAIR_INGREDIENTS = create("tactical_armor_repair_ingredients");
+
+    public static final TagKey<Item> TITANIUM_ARMOR_REPAIR_INGREDIENTS = create("titanium_armor_repair_ingredients");
 
     public static final TagKey<Item> URANIUM_NUGGET_LIKE = create("uranium_nugget_like");
 
@@ -43,7 +53,11 @@ public class HumanItemTags {
 
     public static final TagKey<Item> WY_COMMANDO_ARMOR = create("wy_commando_armor");
 
+    public static final TagKey<Item> WY_COMMANDO_ARMOR_REPAIR_INGREDIENTS = create("wy_commando_armor_repair_ingredients");
+
     public static final TagKey<Item> WY_ELITE_ARMOR = create("wy_elite_armor");
+
+    public static final TagKey<Item> WY_ELITE_ARMOR_REPAIR_INGREDIENTS = create("wy_elite_armor_repair_ingredients");
 
     private static TagKey<Item> create(String name) {
         return TagKey.create(Registries.ITEM, HumanResources.location(name));